copy commit from email freq
