+++ server-side/models/preference.model.js
@@ -4,5 +4,7 @@ const PreferenceSchema= new mongoose.Schema({
     emailFrequency:{type:String,default:'never',enum:['never','daily', 'weekly','monthly','yearly']},
     sendNotificationTime:{type:Number,default:30,required:true},
     soundVoice:{type:String,default:'×××ª ××¢× ×¢×©×'},
+    timeZone:{type:String,default:'UTC'},
+    language:{type:String,default:'en'}
 })
 export default mongoose.model('Preference',PreferenceSchema);
